+++ gradle-plugins/compose/src/main/kotlin/org/jetbrains/compose/experimental/uikit/internal/resources/determineIosKonanTargetsFromEnv.kt
@@ -3,23 +3,9 @@
  * Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE.txt file.
  */
 
-package org.jetbrains.compose.experimental.uikit.internal
+package org.jetbrains.compose.experimental.uikit.internal.resources
 
-import org.jetbrains.kotlin.gradle.plugin.KotlinTarget
-import org.jetbrains.kotlin.gradle.plugin.mpp.KotlinNativeTarget
 import org.jetbrains.kotlin.konan.target.KonanTarget
-
-internal fun KotlinNativeTarget.isIosSimulatorTarget(): Boolean =
-    konanTarget === KonanTarget.IOS_X64 || konanTarget === KonanTarget.IOS_SIMULATOR_ARM64
-
-internal fun KotlinNativeTarget.isIosDeviceTarget(): Boolean =
-    konanTarget === KonanTarget.IOS_ARM64 || konanTarget === KonanTarget.IOS_ARM32
-
-internal fun KotlinNativeTarget.isIosTarget(): Boolean =
-    isIosSimulatorTarget() || isIosDeviceTarget()
-
-internal fun KotlinTarget.asIosNativeTargetOrNull(): KotlinNativeTarget? =
-    (this as? KotlinNativeTarget)?.takeIf { it.isIosTarget() }
 
 // based on AppleSdk.kt from Kotlin Gradle Plugin
 // See https://github.com/JetBrains/kotlin/blob/142421da5b966049b4eab44ce6856eb172cf122a/libraries/tools/kotlin-gradle-plugin/src/common/kotlin/org/jetbrains/kotlin/gradle/plugin/mpp/apple/AppleSdk.kt
