+++ gradle-plugins/compose/src/main/kotlin/org/jetbrains/compose/experimental/uikit/internal/utils/cocoapodsDslHelpers.kt
@@ -3,7 +3,7 @@
  * Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE.txt file.
  */
 
-package org.jetbrains.compose.experimental.uikit.internal
+package org.jetbrains.compose.experimental.uikit.internal.utils
 
 import org.gradle.api.Project
 import org.gradle.api.plugins.ExtensionAware
